Divan.Syllable

Documentation

syllabalize :: Text -> [Syllable]

List of Turkish syllables of the given text. Syllables dont contain apostrophes, they are all in lower case. Note that apostrophe is a definite syllable separator, for words of Arabic origin which containing the sound "ع".

isVowel :: Char -> Bool

Returns True if x is a vowel, or False otherwise

isLongVowel :: Char -> Bool

Returns True if x is a long vowel like 'â', 'ê', 'î', or 'û', or False otherwise

type Syllable = Text

type Sentence = Text