Divan.Tefile

Documentation

type Tefile = Text

Such as "mefâilün"

tefileMap :: [(Symbols, Tefile)]

Association list for vezin symbol strings and tefile names.

tefileLookup :: Vezin -> Maybe Tefile

Tefile name for the given vezin.

tefileSymbolsLookup :: Symbols -> Maybe Tefile

Tefile name for the given Symbols.

inverseLookup :: Tefile -> Maybe Symbols

Symbols for the given tefile.

detectTefile :: Vezin -> Maybe [Tefile]

The shortest tefile list for the vezin.

detectSymbolsTefile :: Symbols -> Maybe [Tefile]

Find the shortest tefile list for the Symbols by breadth-first search. Note that we have a preference to find the longest possible tefile name at a time because we don't want to end up with a tefile list with one syllable names.

tefileName :: [Tefile] -> Text

Returns a string containing all tefiles separated by /

equivalent :: [Tefile] -> [Tefile] -> Bool

A function to determine if two tefile lists correspond to the same vezin. This function is necesssary because this program tries to find the longest tefile names that fit the vezin. Sometimes you might be looking to see if a tefile list fits a verse, this is the function to use then.